Performance Measure, Weights, & Better Comparisons
Created a measure of performance. This is the number of matches in the topMatches list that are known to be matches (matching unique identifier). 

Implemented weights per field. Currently, default weights are implemented that achieve ~78% match accuracy.

Also changed the way comparisons are handled. This new method is much more flexible and allows rules to be placed for each field.