New Data, Parallel Processing, & More!
We've obtained another set of data for matching to the master record. This new dataset seems to have more issues, but alas, we've managed to hold greater than 90% accuracy of known matches.

Also converted our loop structure to support parallel streaming. This took us from ~25 minutes to match ~20,000 x ~20,000, down to 8-10 minutes.

Implemented an alternative set of weights to use given missing email fields. This brought our accuracy on the original matching dataset from 93% to 99%, and our new matching dataset from 60% to 93%.

Implemented a hashmap for common address abbreviations. During the pre-processing stage, the address fields are combined, then sent the the PreProcessor object. The address string is split by whitespace, capitalized, and then checks each word to see if it's an abbreviation. If it is, then it gets changed to it's full-form counterpart.

Some other things were changed, but this covers this most important & beneficial changes.