Refactoring & New Features
CSV headers were standardized to give CSVReader a quality of life improvement. Now uses a private enum to store the header values. This lead to an increase in accuracy with a *slight* (0.01) decrease in comparison reductions. This was likely due to an uncaught error in our previous method of reading the matching CSV files.

MatchRunner class was created to de-clutter our main method. This class also has functionality to grab all files in the path to locate matching CSV's. This is set in config/DataMatching.properties

Small changes to other classes. Mostly tweaking them to work with the CSVReader refactor.