Javadoc Fix & CSV Input Errors
Last push I refactored the 'data' package to 'dataholder'. This found every instance of the word data and replaced it...mostly in the java docs. Changed all instances of this in the java docs back to the word 'data'.

Fixed masterToMaster bug in MatchRunner.

Changed readMaster and readMatch functions to return void & they now throw custom exceptions that will error out if a CSV file fails to be found / read.